neutrino-image-base.inc: synch 'imageversion' file with release text file
Formally synchronize the generation of the 'imageversion' file with
the release text file (e.g., 'beta.txt'), which indicates the latest
created image version. The 'imageversion' file is already generated
during the image creation process, along with the kernel and the rootfs,
but only in the machine layers (e.g., 'IMAGE_CMD_airdigitalemmc_append').

To avoid repeating this in every machine layer, it is easier to do it only
once so that it is available everywhere. Since the image name is already
determined, any timestamps or other features contained in it should be
identical, ensuring no unexpected differences arise.